This commit applies all of the post-dev changes to model directives:
 * Sync with upstream changes
 * Rewrite to not reference dev
 * Removing invalid geometry from the jaco (as mentioning
   it in a data= line causes it to be validity checked)
 * Factor the scoped names mechanism out of model directives
 * Rewrite the weld error API to be future-proof against a
   future refactor.
