+++ multibody/parsing/README_model_directives.md
@@ -1,41 +1,29 @@
 The Model Directives mechanism
 ==============================
 
-Model Directives is a small yaml-based language for building a complex
-MultibodyPlant-based scene out of numerous SDFs.  For instance in the TRI
-dish-loading demo we have individual SDF files for the counter, sink, cameras,
-pedestal, arm, gripper, and each manipuland.  A single SDF for this would be
-unwieldy and difficult to maintain and collaborate on, but SDF's file
-inclusion mechanisms have not yet proven adequate to this task.
+"Model Directives" is a small YAML-based language for building a complex
+MultibodyPlant-based scene out of numerous `.sdf` and `.urdf` files.  For
+instance in the TRI dish-loading demo we have individual SDFormat files for
+the counter, sink, cameras, pedestal, arm, gripper, and each manipuland.  A
+single SDFormat for this would be unwieldy and difficult to maintain and
+collaborate on, but SDFormat's file inclusion mechanisms have not yet proven
+adequate to this task.
 
-We expect that this mechanism will be temporary and will be removed when
-sdformat adds similar functionality.  Users should be aware that this library
-will be deprecated if/when sdformat reaches feature parity with it.
+We expect that this mechanism will be temporary and will be removed
+[when SDFormat adds similar functionality](#conditions-for-deprecation).
+Users should be aware that this library will be deprecated if/when sdformat
+reaches feature parity with it.
 
 
 ## Syntax
 
-The easiest syntax reference is the unit test files in `test/models/*.yaml` of
-this directory.
+The easiest syntax reference is the unit test files in
+`test/process_model_directives_test/*.yaml` of this directory.
 
-A model directives file is a yaml file with a top level `directives:` group.
-Within this group are a series of directives:
-
- * `AddModel` takes a `file` and `name` and loads the SDF/URDF file indicated
-   as a new model instance with the given name.
- * `AddModelInstance` Creates a new, empty model instance in the plant with
-   the indicated `name`.
- * `AddPackagePath` takes `name` and `path` and makes `package://name` URIs
-   be resolved to `path`.  This directive is due for deprecation soon and
-   should generally be avoided.
- * `AddFrame` takes a `name` and a `X_PF` transform and adds a new frame to
-   the model.  Note that the transform is as specified in the `Transform`
-   scenario schema and can reference an optional base frame, in which case
-   the frame will be added to the base frame's model instance.
- * `AddDirectives` takes a `file` naming a model directives file and an
-   optional `model_namespace`; it loads the model directives from that file
-   with the namespace prefixed to them (see Scoping, below).
- * `AddWeld` takes a `parent` and `child` frame and welds them together.
+A model directives file is a YAML file with a top level `directives:` group.
+Within this group are a series of directives, which are described in detail in
+`model_directives.h`.  These directives can add model instances and add frames
+to them and welds between them.
 
 
 ## Use
@@ -52,6 +40,18 @@ A simple example of a use would be:
   plant.Finalize();
 ```
 
+Or, if you are connecting a scene graph:
+
+```cpp
+  DiagramBuilder builder;
+  ModelDirectives station_directives = LoadModelDirectives(
+      FindResourceOrThrow("my_exciting_project/models/my_scene.yaml"));
+  auto [plant, scene_graph] =
+      AddMultibodyPlantSceneGraph<double>(&builder, 0.0);;
+  ProcessModelDirectives(station_directives, &plant);
+  plant.Finalize();
+```
+
 This loads the model directives from that filename, constructs a plant, and
 uses the model directives to populate the plant.
 
@@ -59,31 +59,45 @@ uses the model directives to populate the plant.
 ## Scoping
 
 Elements (frames, bodies, etc.) in `MultibodyPlant` belong to model instances.
-Model instances can have any name specifiers, and can contain the "namespace"
-delimiter `::`. Element names should not contain `::`.
+To acknowledge this, and support "faux" composition, we introduce scoping via
+names.
 
-Examples:
+Names can be scoped implicitly or explicitly, and this scoping supports
+multiple levels of "nesting" -- insofar as model instances in MultibodyPlant
+can denote hierarchy (see #14043 for more info).
 
-- `my_frame` implies no explicit model instance.
-- `my_model::my_frame` implies the model instance `my_model`, the frame
-`my_frame`.
-- `top_level::my_model::my_frame` implies the model instance
-`top_level::my_model`, the frame `my_frame`.
+Names are delimited by `::`; the last name portion indicates the element name,
+and all preceding portions indicate the model instance name. For this reason,
+element names themselves should not contain `::`.
+
+(The `::` convention is used for consistency with the emerging namespace
+syntax for SDFormat, soon to be required in SDFormat 1.8.)
+
+Thus:
+
+- `my_frame`: No explicit model instance, the frame is `my_frame`.
+- `my_model::my_frame`: The model instance is `my_model`, the frame is
+  `my_frame`.
+- `top_level::my_model::my_frame`: The model instance is
+  `top_level::my_model`, the frame is `my_frame`.
 
 
 ## Conditions for deprecation
 
 We expect and hope to deprecate this mechanism when either:
 
-SDF format properly specifies, and Drake supports, the following:
+SDFormat properly specifies, and Drake supports, the following:
 
  * What `<include/>` statements should *really* do (e.g. namespacing models,
    joints, etc.) without kludging Drake's parsing
+   * See <http://sdformat.org/tutorials?tut=composition_proposal&cat=pose_semantics_docs&> for the progress of the proposed formal semantics.
  * How to weld models together with joints external to the models
+
 
 OR if we find a mechanism whereby xacro could accomplish the same thing:
 
  * Drake's `package://` / `model://` mechanism were mature and correct, and
    `sdformat` didn't use singletons for search paths.
  * It was easier for one xacro to locate other xacros (possibly via a
-   workaround using a wrapper script to inject `DRAKE_PATH`)
+   workaround using a wrapper script to inject the appropriate drake source
+   or resource path)
