+++ multibody/parsing/test/process_model_directives_test/simple_model.sdf
@@ -17,6 +17,7 @@
               package:// URI resolution.
           -->
           <drake:diffuse_map>
+            <!-- TODO((#10531) This should use `package://drake`. -->
             package://process_model_directives_test/circle.png
           </drake:diffuse_map>
         </material>
